fix xss attack

+++ backend/src/main/java/com/dev/backend/service/PostService.java
@@ -9,6 +9,8 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -25,7 +27,6 @@ import com.dev.backend.dto.DetailPostResponse;
 import com.dev.backend.dto.DiscoveryPostResponse;
 import com.dev.backend.dto.EditPostResponse;
 import com.dev.backend.dto.FeedPostResponse;
-import com.dev.backend.dto.PostRequest;
 import com.dev.backend.dto.ProfilePostResponse;
 import com.dev.backend.dto.UploadResponse;
 import com.dev.backend.dto.UserDto;
@@ -73,25 +74,22 @@ public class PostService {
         this.userRepository = userRepository;
     }
 
-    public Post savePost(PostRequest postDto, UUID userId) throws SafeHtmlException {
-        String sanitizedTitle = htmlSanitizerService.sanitizeTitle(postDto.title());
-        String sanitizedContent = htmlSanitizerService.sanitizeContent(postDto.content());
+    public Post savePost(String title, String content, MultipartFile thumbnail, List<MultipartFile> files, UUID userId)
+            throws SafeHtmlException, IOException {
 
-        if (sanitizedTitle.trim().isEmpty()) {
-            throw new SafeHtmlException("Title cannot be empty or contain only HTML");
-        }
+        UploadResponse uploadResponse = upload(thumbnail, files);
 
-        if (sanitizedContent.trim().isEmpty()) {
-            throw new SafeHtmlException("Title cannot be empty or contain only HTML");
-        }
+        content = addLinkToSrc(content, uploadResponse.fileNames());
+
+        String sanitizeContent = htmlSanitizerService.sanitizeContent(content);
 
         User user = userRepository.getReferenceById(userId);
         Post post = new Post();
-        post.setTitle(sanitizedTitle);
-        post.setContent(postDto.content());
-        post.setFiles(String.join(", ", postDto.files()));
+        post.setTitle(title);
+        post.setContent(sanitizeContent);
+        post.setFiles(String.join(", ", uploadResponse.fileNames()));
         post.setUser(user);
-        post.setThumbnail(postDto.thumbnail());
+        post.setThumbnail(uploadResponse.thumbnail());
         post.setStatus(PostStatus.PUBLISHED);
         postRepository.save(post);
         List<Follow> followers = user.getFollowers();
@@ -103,27 +101,64 @@ public class PostService {
         return post;
     }
 
-    public Post updatePost(UUID id, PostRequest updatedPost, UUID currentUserId) throws SafeHtmlException {
-        String sanitizedTitle = htmlSanitizerService.sanitizeTitle(updatedPost.title());
-        String sanitizedContent = htmlSanitizerService.sanitizeContent(updatedPost.content());
-
-        if (sanitizedTitle.trim().isEmpty()) {
-            throw new SafeHtmlException("Title cannot be empty or contain only HTML");
-        }
-
-        if (sanitizedContent.trim().isEmpty()) {
-            throw new SafeHtmlException("Title cannot be empty or contain only HTML");
-        }
+    public Post updatePost(UUID id, String title, String content, MultipartFile thumbnail, List<MultipartFile> files,
+            String oldThumbnail, List<String> oldFileNames,
+            UUID currentUserId)
+            throws SafeHtmlException, IOException {
 
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Post not found"));
+
         if (!post.getUser().getId().equals(currentUserId)) {
             throw new AccessDeniedException("You cannot delete another user's post.");
         }
-        post.setTitle(sanitizedTitle);
-        post.setContent(updatedPost.content());
-        post.setFiles(String.join(", ", updatedPost.files()));
-        post.setThumbnail(updatedPost.thumbnail());
+
+        String finalThumbnail;
+        if (thumbnail != null && !thumbnail.isEmpty()) {
+            finalThumbnail = upload(thumbnail, null).thumbnail();
+            deleteMedia(oldThumbnail, "");
+        } else {
+            finalThumbnail = oldThumbnail.substring(oldThumbnail.lastIndexOf('/') + 1);
+        }
+
+        List<String> finalFileNames = new ArrayList<>();
+        if (oldFileNames != null && !oldFileNames.isEmpty()) {
+            List<String> newUploadedFiles = new ArrayList<>();
+            if (files != null && !files.isEmpty()) {
+                newUploadedFiles = upload(null, files).fileNames();
+            }
+
+            int newFileIndex = 0;
+            for (String fileName : oldFileNames) {
+                if (fileName.startsWith("new_file_")) {
+                    finalFileNames.add(newUploadedFiles.get(newFileIndex++));
+                } else {
+                    finalFileNames.add(fileName);
+                }
+            }
+        }
+
+        content = addLinkToSrc(content, finalFileNames);
+
+        String sanitizeContent = htmlSanitizerService.sanitizeContent(content);
+
+        if (!post.getFiles().equals("")) {
+            String[] filesToCheck = post.getFiles().split(", ");
+            List<String> filesToDelete = new ArrayList<>();
+            for (String file : filesToCheck) {
+                if (!sanitizeContent.contains(file)) {
+                    filesToDelete.add(file);
+                }
+            }
+            if (!filesToDelete.isEmpty()) {
+                deleteMedia(null, String.join(", ", filesToDelete));
+            }
+        }
+
+        post.setTitle(title);
+        post.setContent(sanitizeContent);
+        post.setFiles(String.join(", ", finalFileNames));
+        post.setThumbnail(finalThumbnail);
         return postRepository.save(post);
     }
 
@@ -305,16 +340,18 @@ public class PostService {
 
     public UploadResponse upload(MultipartFile thumbnail, List<MultipartFile> files) throws IOException {
         String thumbnailId = "";
-        String thumbnailName = thumbnail.getOriginalFilename();
-        String thumbnailExtension = (thumbnailName != null && thumbnailName.contains("."))
-                ? thumbnailName.substring(thumbnailName.lastIndexOf("."))
-                : "";
-        thumbnailId = UUID.randomUUID() + thumbnailExtension;
+        if (thumbnail != null) {
+            String thumbnailName = thumbnail.getOriginalFilename();
+            String thumbnailExtension = (thumbnailName != null && thumbnailName.contains("."))
+                    ? thumbnailName.substring(thumbnailName.lastIndexOf("."))
+                    : "";
+            thumbnailId = UUID.randomUUID() + thumbnailExtension;
 
-        Path path = Paths.get(uploadDir + "/images", thumbnailId);
+            Path path = Paths.get(uploadDir + "/images", thumbnailId);
 
-        Files.createDirectories(path.getParent());
-        Files.copy(thumbnail.getInputStream(), path, StandardCopyOption.REPLACE_EXISTING);
+            Files.createDirectories(path.getParent());
+            Files.copy(thumbnail.getInputStream(), path, StandardCopyOption.REPLACE_EXISTING);
+        }
         List<String> fileNames = new ArrayList<>();
         if (files != null && !files.isEmpty()) {
             for (MultipartFile file : files) {
@@ -330,7 +367,7 @@ public class PostService {
                     subDir = (contentType.startsWith("image")) ? "/images" : "/videos";
                 }
 
-                path = Paths.get(uploadDir + subDir, id);
+                Path path = Paths.get(uploadDir + subDir, id);
 
                 Files.createDirectories(path.getParent());
                 Files.copy(file.getInputStream(), path, StandardCopyOption.REPLACE_EXISTING);
@@ -364,13 +401,15 @@ public class PostService {
     }
 
     private void deleteMedia(String thumbnail, String files) throws IOException {
-        Path pathThumbnail = Paths.get(uploadDir + "/images").resolve(thumbnail).normalize();
+        if (thumbnail != null) {
+            Path pathThumbnail = Paths.get(uploadDir + "/images").resolve(thumbnail).normalize();
 
-        if (pathThumbnail == null || !Files.exists(pathThumbnail)) {
-            return;
+            if (pathThumbnail == null || !Files.exists(pathThumbnail)) {
+                return;
+            }
+
+            Files.delete(pathThumbnail);
         }
-
-        Files.delete(pathThumbnail);
 
         if (files.equals(""))
             return;
@@ -388,5 +427,22 @@ public class PostService {
 
             Files.delete(filePath);
         }
+    }
+
+    private String addLinkToSrc(String htmlString, List<String> fileNames) {
+        Pattern pattern = Pattern.compile("(data:(?:image|video)/[a-zA-Z0-9+.-]+;base64,[^\\\\\\\"']+)");
+        Matcher matcher = pattern.matcher(htmlString);
+        StringBuffer result = new StringBuffer();
+        int index = 0;
+
+        while (matcher.find()) {
+            if (index < fileNames.size()) {
+                String newUrl = "http://localhost:8080/api/post/file/" + fileNames.get(index++);
+                String replacement = newUrl;
+                matcher.appendReplacement(result, Matcher.quoteReplacement(replacement));
+            }
+        }
+        matcher.appendTail(result);
+        return result.toString();
     }
 }
